Changed map based board into array based
I initially write the board initial state as Board type which is a mapped object that contains column and array of task. I encountered an error using zustand persist middleware and decided to change the board initial state as an array and then gladly the persist works with a board state that is array based.